refactor(atlassian): move attachment_filename to utils::path
Move the safe single-component filename helper (and its unit tests)
from the Confluence bulk-download module to a new crate-internal
src/utils/path.rs so the other attachment-download sites can reuse
it. No behavior change; adds one test covering absolute-path titles.

Preparation for the #1127 path-traversal fix.
